unify workspace and single-contract build logic
